removed models, public, routes, and views, folders and the listly.html file, and added working public, and routes folders, as well as the schema.sql file
