--- public/assets/list.html
@@ -1,41 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <title>Listly!</title>
-  <!-- Latest compiled and minified CSS & JS -->
-  <link href="https://maxcdn.bootstrapcdn.com/bootswatch/4.0.0/flatly/bootstrap.min.css" rel="stylesheet">
-  <link rel="stylesheet" href="./css/style.css" media="screen" title="no title">
-  <script src="https://code.jquery.com/jquery.js"></script>
-  <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"></script>
-</head>
-
-<body>
-  <div class="container">
-    <div class="row mt-4">
-      <div class="col-md-4">
-        <div class="form-group">
-          <select class="custom-select" id="category">
-            <option selected value="">All Categories</option>
-            <option value="Personal">Personal</option>
-            <option value="Work">Work</option>
-          </select>
-        </div>
-      </div>
-      <div class="col-md-2">
-        <a class="btn btn-light" href="/to-do">New To Do</a>
-      </div>
-    </div>
-    <div class="row">
-      <div class="col-md-12 blog-container">
-      </div>
-    </div>
-  </div>
-  <!-- Custom Script -->
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.17.0/moment.min.js" type="text/javascript"></script>
-  <script src="./js/script.js" type="text/javascript"></script>
-
-</body>
-
-</html>
