+++ public/list.html
@@ -1,0 +1,80 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <title>To-Do</title>
+  <!-- Latest compiled and minified CSS & JS -->
+  <link href="https://maxcdn.bootstrapcdn.com/bootswatch/4.0.0/flatly/bootstrap.min.css" rel="stylesheet">
+  <link rel="stylesheet" href="css/styles.css" media="screen" title="no title">
+  <script src="https://code.jquery.com/jquery.js"></script>
+  <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"></script>
+</head>
+
+<body class="homepage">
+  <div class="navbar">
+    <a class="active" href="./list.html"><i class="fa fa-fw fa-list"></i> List</a>
+    <a class="other" href="./todo.html"><i class="fa fa-fw fa-todo"></i> To-do</a>
+    <div class="row mt-12 task-bar">
+      <div class="col-md-6">
+        <div class="form-group">
+          <select class="custom-select" id="urgency">
+            <option selected value="">All Categories</option>
+            <option value="low">Low</option>
+            <option value="medium">Medium</option>
+            <option value="high">High</option>
+          </select>
+        </div>
+      </div>
+      <div class="col-md-6">
+        <a class="btn  new-task " href="/todo">New Task</a>
+      </div>
+    </div>
+  </div>
+
+
+  <div class="container">
+    <div class="center">
+      <div class="row">
+        <div class="col-12">
+          <!-- <img src="/public/img/Start Page Background.jpg" alt="list-icon" class="center list-icon"> -->
+        </div>
+        <div class="row">
+          <div class="col-12">
+            <!-- <h1 class="getstarted"> Need to add a task?<br>Let's get started.</h1> -->
+          </div>
+        </div>
+        <div class="container tast-bar">
+          <!-- <div class="row mt-12">
+            <div class="col-md-6">
+              <div class="form-group">
+                <select class="custom-select" id="urgency">
+                  <option selected value="">All Categories</option>
+                  <option value="low">Low</option>
+                  <option value="medium">Medium</option>
+                  <option value="high">High</option>
+                </select>
+              </div>
+            </div>
+            <div class="col-md-6">
+              <a class="btn  new-task " href="/todo">New Task</a>
+            </div>
+          </div> -->
+          <div class="row">
+            <div class="col-md-12 list-container">
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+
+
+  <!-- Custom Script -->
+
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.17.0/moment.min.js" type="text/javascript"></script>
+  <script src="js/list.js" type="text/javascript"></script>
+
+</body>
+
+</html>
